Move `CoreDataFeedStore` implementation of `FeedStore` protocol into a separate file
